# swagger.json
{"openapi":"3.0.0","info":{"title":"Node S15","version":"1.0.0","description":"This is a simple CRUD API","license":{"name":"MIT","url":"http://mit.com"},"contact":{"name":"Fran Linde","url":"https://github.com/franlidebl","email":"fran@example.com"}},"server":[{"url":"http://localhost:3000"}],"paths":{"/author":{"get":{"summary":"Get a list of authors","tags":["Authors"],"parameters":[{"in":"query","name":"page","schema":{"type":"integer"},"description":"Page number"},{"in":"query","name":"limit","schema":{"type":"integer"},"description":"Number of items per page"}],"responses":{"200":{"description":"A list of authors","content":{"application/json":{"schema":{"type":"object","properties":{"totalItems":{"type":"integer","description":"Total number of items"},"totalPages":{"type":"integer","description":"Total number of pages"},"currentPage":{"type":"integer","description":"Current page number"},"data":{"type":"array","items":{"$ref":"#/components/schemas/Author"}}}}}}},"500":{"description":"Internal server error"}}},"post":{"summary":"Create a new author","tags":["Authors"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Author"}}}},"responses":{"201":{"description":"The created author","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Author"}}}},"400":{"description":"Invalid request body parameters","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}}}}}},"500":{"description":"Internal server error"}}}},"/author/{id}":{"get":{"summary":"Get an author by ID","tags":["Authors"],"parameters":[{"in":"path","name":"id","schema":{"type":"string"},"required":true,"description":"Author ID"}],"responses":{"200":{"description":"The author","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Author"}}}},"404":{"description":"Author not found"},"500":{"description":"Internal server error"}}},"delete":{"summary":"Delete an author by ID","tags":["Authors"],"parameters":[{"in":"path","name":"id","schema":{"type":"string"},"required":true,"description":"Author ID"}],"responses":{"200":{"description":"The deleted author","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Author"}}}},"401":{"description":"Unauthorized operation"},"404":{"description":"Author not found"},"500":{"description":"Internal server error"}}},"put":{"summary":"Update an author by ID","tags":["Authors"],"parameters":[{"in":"path","name":"id","schema":{"type":"string"},"required":true,"description":"Author ID"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Author"}}}},"responses":{"200":{"description":"The updated author","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Author"}}}},"400":{"description":"Invalid request body parameters","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}}}}}},"401":{"description":"Unauthorized operation"},"404":{"description":"Author not found"},"500":{"description":"Internal server error"}}}},"/author/name/{name}":{"get":{"summary":"Get authors by name","tags":["Authors"],"parameters":[{"in":"path","name":"name","schema":{"type":"string"},"required":true,"description":"Author name"}],"responses":{"200":{"description":"A list of authors","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Author"}}}}},"404":{"description":"Authors not found"},"500":{"description":"Internal server error"}}}},"/author/image-upload":{"post":{"summary":"Upload an author's image","tags":["Authors"],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"logo":{"type":"string","format":"binary"},"authorId":{"type":"string"}}}}}},"responses":{"200":{"description":"The updated author with the new image","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Author"}}}},"400":{"description":"Invalid request body parameters"},"404":{"description":"Author not found"},"500":{"description":"Internal server error"}}}},"/author/login":{"post":{"summary":"User login","tags":["Authors"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}}}}},"responses":{"200":{"description":"User login successful","content":{"application/json":{"schema":{"type":"object","properties":{"token":{"type":"string"}}}}}},"400":{"description":"Invalid request body parameters"},"401":{"description":"Unauthorized access"},"500":{"description":"Internal server error"}}}},"/book":{"get":{"summary":"Get all books","tags":["Books"],"parameters":[{"in":"query","name":"page","schema":{"type":"integer","minimum":1},"description":"Page number"},{"in":"query","name":"limit","schema":{"type":"integer","minimum":1},"description":"Number of books per page"}],"responses":{"200":{"description":"A list of books","content":{"application/json":{"schema":{"type":"object","properties":{"totalItems":{"type":"integer","description":"Total number of books"},"totalPages":{"type":"integer","description":"Total number of pages"},"currentPage":{"type":"integer","description":"Current page number"},"data":{"type":"array","items":{"$ref":"#/components/schemas/Book"}}}}}}},"400":{"description":"Invalid page or limit parameters"},"500":{"description":"Internal server error"}}},"post":{"summary":"Create a new book","tags":["Books"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Book"}}}},"responses":{"201":{"description":"The created book","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Book"}}}},"400":{"description":"Invalid request body parameters"},"500":{"description":"Internal server error"}}}},"/book/{id}":{"get":{"summary":"Get a book by ID","tags":["Books"],"parameters":[{"in":"path","name":"id","schema":{"type":"string"},"required":true,"description":"Book ID"}],"responses":{"200":{"description":"The book with the specified ID","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Book"}}}},"404":{"description":"Book not found"},"500":{"description":"Internal server error"}}},"delete":{"summary":"Delete a book by ID","tags":["Books"],"parameters":[{"in":"path","name":"id","schema":{"type":"string"},"required":true,"description":"Book ID"}],"responses":{"200":{"description":"The deleted book","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Book"}}}},"404":{"description":"Book not found"},"500":{"description":"Internal server error"}}},"put":{"summary":"Update a book by ID","tags":["Books"],"parameters":[{"in":"path","name":"id","schema":{"type":"string"},"required":true,"description":"Book ID"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Book"}}}},"responses":{"200":{"description":"The updated book","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Book"}}}},"400":{"description":"Invalid request body parameters"},"404":{"description":"Book not found"},"500":{"description":"Internal server error"}}}},"/book/brand/{brand}":{"get":{"summary":"Get books by brand","tags":["Books"],"parameters":[{"in":"path","name":"brand","schema":{"type":"string"},"required":true,"description":"Brand name"}],"responses":{"200":{"description":"Books with the specified brand","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Book"}}}}},"404":{"description":"Books not found"},"500":{"description":"Internal server error"}}}}},"components":{"schemas":{"Author":{"type":"object","properties":{"email":{"type":"string","format":"email","description":"The email address of the author."},"password":{"type":"string","minLength":8,"description":"The password of the author."},"name":{"type":"string","minLength":3,"maxLength":25,"description":"The name of the author."},"country":{"type":"string","enum":["SPAIN","ITALY","USA","GERMANY","JAPAN","FRANCE","ENGLAND","COLOMBIA","RUSSIA","ARGENTINA","CZECHOSLOVAKIA","NIGERIA"],"description":"The country of the author."},"profileImage":{"type":"string","description":"The URL of the author's profile image."}},"required":["email","password","name","country"]},"Book":{"type":"object","properties":{"title":{"type":"string","minLength":3,"maxLength":50,"description":"The title of the book."},"author":{"type":"string","description":"The ID of the author of the book."},"pages":{"type":"number","minimum":1,"maximum":10000,"description":"The number of pages in the book."},"publisher":{"type":"object","properties":{"name":{"type":"string","minLength":3,"maxLength":25,"description":"The name of the publisher."},"country":{"type":"string","enum":["SPAIN","ITALY","USA","GERMANY","JAPAN","FRANCE"],"description":"The country of the publisher."}}}},"required":["title","pages","publisher"]}}},"tags":[{"name":"Authors","description":"API endpoints for authors"},{"name":"Books","description":"API endpoints for books"}]}